+++ src/server.ts
@@ -1,72 +1,26 @@
 import './config/env';
 import 'reflect-metadata';
 import './typeorm';
-import './config/env';
 import { ApolloServer, gql } from 'apollo-server';
-import { getRepository } from 'typeorm';
-import { compare } from 'bcryptjs';
-import User from './typeorm/entities/User';
+import { userResolvers, userTypeDefs } from './typeorm/entities/users/Graphql';
 
-const typeDefs = gql`
-  type User {
-    id: String!
-    name: String!
-    email: String!
-    birthDate: String!
-    cpf: String!
-  }
-
-  type LoginResponse {
-    user: User!
-    token: String!
-  }
-
+const helloTypeDefs = gql`
   type Query {
     hello: String
   }
-
-  type Mutation {
-    login(login: LoginInput!): LoginResponse!
-  }
-
-  input LoginInput {
-    "User email"
-    email: String!
-    "User password"
-    password: String!
-  }
 `;
 
-const resolvers = {
+const helloResolvers = {
   Query: {
     hello: () => {
       return 'hello, world!';
     },
   },
-
-  Mutation: {
-    login: async (_, { login }) => {
-      const { email, password } = login;
-      const usersRepository = getRepository(User);
-
-      if (!email || !password) {
-        throw new Error('Unauthorized. Possible invalid credentials.');
-      }
-
-      const user = await usersRepository.findOne({ email });
-
-      const passswordMatch = await compare(password, user.password);
-
-      if (!user || !passswordMatch) {
-        throw new Error('Unauthorized. Possible invalid credentials.');
-      }
-
-      const token = 'Token';
-
-      return { user, token };
-    },
-  },
 };
+
+const typeDefs = [helloTypeDefs, userTypeDefs];
+
+const resolvers = [helloResolvers, userResolvers];
 
 const server = new ApolloServer({ typeDefs, resolvers });
